requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.5
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
idna==3.11
psycopg2-binary==2.9.11
PyMySQL==1.1.2
python-dotenv==1.0.0
pytz==2026.1.post1
requests==2.31.0
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
